# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ipe Distribuidora API
  version: "1.0.0"
servers:
  - url: http://localhost:3000
paths:
...
  /api/brands/{id}:
    get:
      summary: Get brand by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    patch:
      summary: Update brand
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandUpdateRequest"
      responses:
        "200":
          description: OK
    delete:
      summary: Delete brand
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...
  /api/products/{id}:
    get:
      summary: Get product by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    patch:
      summary: Update product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdateRequest"
      responses:
        "200":
          description: OK
    delete:
      summary: Soft delete product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...
components:
  schemas:
...
    BrandUpdateRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
...
    ProductUpdateRequest:
      type: object
      properties:
        sku:
          type: string
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string
          nullable: true
        isActive:
          type: boolean
...